API

Types

BEA.AMNE — Type

AMNE(
    directionofinvestment::AbstractString,
    allaffiliates::Bool,
    nonbankaffiliatesonly::Bool,
    classification::AbstractString,
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}};
    seriesid::Union{AbstractString, Integer, AbstractVector{<:Integer}} = "ALL",
    country::Union{AbstractString, Integer, AbstractVector{<:Integer}} = "ALL",
    industry::Union{AbstractString, AbstractVector{<:AbstractString}} = "ALL",
    state::AbstractString = "ALL",
    getfootnotes::Bool = false
    ) -> AMNE

Activities of Multinational Enterprises (AMNE) — operations and finances of U. S. parent enterprises and their foreign affiliates and U. S. affiliates of foreign MNEs.

Examples

julia> query = AMNE("Outward", false, false, "CountryByIndustry", 2011:2012, seriesid = 4:5, country = 202)
AMNE("Outward", "0", "0", "4,5", "CountryByIndustry", "2011,2012", "202", "ALL", "ALL", "No")

julia> query = AMNE("Inward", false, false, "Country", 2011, seriesid = 8, industry = "0000")
AMNE("Inward", "0", "0", "8", "Country", "2011", "ALL", "0000", "ALL", "No")

BEA.DirectInvestment — Type

DirectInvestment(
    directionofinvestment::AbstractString,
    classification::AbstractString,
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}};
    country::Union{AbstractString, Integer, AbstractVector{<:Integer}} = "ALL",
    industry::Union{AbstractString, AbstractVector{<:AbstractString}} = "ALL",
    seriesid::Union{AbstractString, Integer, AbstractVector{<:Integer}} = "ALL",
    getfootnotes::Bool = false
    ) -> DirectInvestment

Direct Investment (DI) — income and financial transactions in direct investment that underlie the U. S. balance of payments statistics, and direct investment positions that underlie the U. S. international investment positions.

Examples

julia> query = DirectInvestment("Outward", "Country", 2011:2012, country = [650, 699])
DirectInvestment("Outward", "ALL", "Country", "2011,2012", "650,699", "ALL", "No")

julia> query = DirectInvestment("Inward", "CountryByIndustry", 2010:2013, country = 308, industry = "3000")
DirectInvestment("Inward", "ALL", "CountryByIndustry", "2010,2011,2012,2013", "308", "3000", "No")

BEA.FixedAssets — Type

FixedAssets(
    tablename::AbstractString,
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}}
    ) -> FixedAssets

This dataset contains data from the standard set of Fixed Assets tables as published online.

Examples

julia> query = FixedAssets("FAAt201", "ALL")
FixedAssets("FAAt201", "ALL")

julia> query = FixedAssets("FAAt405", 2015:2016)
FixedAssets("FAAt405", "2015,2016")

BEA.GDPbyIndustry — Type

GDPbyIndustry(
    tableid::Union{AbstractString, Integer},
    frequency::Union{Char, AbstractVector{<:Char}},
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}},
    industry::Union{AbstractString, AbstractVector{<:AbstractString}}
    ) -> GDPbyIndustry

The gross domestic product by industry data are contained within a dataset called GDPbyIndustry. BEA's industry accounts are used extensively by policymakers and businesses to understand industry interactions, productivity trends, and the changing structure of the U. S. economy. The GDP-by-industry dataset includes data in both current and chained (real) dollars. The dataset contains estimates for value added, gross output, intermediate inputs, KLEMS and employment statistics.

Examples

julia> query = GDPbyIndustry(1, 'A', 2011:2012, "ALL")
GDPbyIndustry("1", "A", "2011,2012", "ALL")

julia> query = GDPbyIndustry(11, ['A', 'Q'], 2010, "11")
GDPbyIndustry("11", "A,Q", "2010", "11")

BEA.IIP — Type

IIP(;
    typeofinvestment::Union{AbstractVector{<:AbstractString}, AbstractString} = "ALL",
    component::Union{AbstractVector{<:AbstractString}, AbstractString} = "ALL",
    frequency::Union{Char, AbstractVector{<:Char}} = ['A', 'N'],
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}} = "ALL"
    ) -> IPP

This dataset contains data on the U. S. international investment position. BEA's international investment position accounts include the end of period value of accumulated stocks of U. S. financial assets and liabilities.

Examples

julia> query = IIP(typeofinvestment = "FinAssetsExclFinDeriv", component = "ChgPosPrice", frequency = 'A')
IIP("FinAssetsExclFinDeriv", "ChgPosPrice", "A", "ALL")

julia> query = IIP(typeofinvestment = "FinLiabsFoa", component = "Pos", frequency = 'N', year = 2013)
IIP("FinLiabsFoa", "Pos", "QNSA", "2013")

BEA.ITA — Type

ITA(
    indicator::Union{AbstractVector{<:AbstractString}, AbstractString} = "ALL",
    areaorcountry::Union{AbstractVector{<:AbstractString}, AbstractString} = "ALL",
    frequency::Union{Char, AbstractVector{<:Char}} = ['A', 'S', 'N'],
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}} = "ALL"
    ) -> ITA

International Transactions data on U.S. international transactions. BEA's international transactions (balance of payments) accounts include all transactions between U. S. and foreign residents.

Examples

julia> query = ITA(indicator = "BalGds", areaorcountry = "China", frequency = 'A', year = 2011:2012)
ITA("BalGds", "China", "A", "2011,2012")

julia> query = ITA(indicator = "PfInvAssets", areaorcountry = "AllCountries", frequency = 'N', year = 2013)
ITA("PfInvAssets", "AllCountries", "QNSA", "2013")

BEA.InputOutput — Type

InputOutput(
    tableid::Union{Integer, AbstractVector{<:Integer}},
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}}
    ) -> InputOutput

The input-output accounts provide a detailed view of the interrelationships between U.S. producers and users. The Input‐Output dataset contains Make Tables, Use Tables, and Direct and Total Requirements tables.y. The input-output accounts provide a detailed view of the interrelationships between U.S. producers and users. The Input‐Output dataset contains Make Tables, Use Tables, and Direct and Total Requirements tables.

Examples

julia> query = InputOutput(56, 2010:2013)
InputOutput("56", "2010,2011,2012,2013")

julia> query = InputOutput(56:57, 2007)
InputOutput("56,57", "2007")

BEA.IntlServTrade — Type

IntlServTrade(;
    typeofservice::Union{AbstractString, AbstractVector{<:AbstractString}} = "ALL",
    tradedirection::Union{AbstractString, AbstractVector{<:AbstractString}} = "ALL",
    affiliation::Union{AbstractString, AbstractVector{<:AbstractString}} = "ALL",
    areaorcountry::Union{AbstractString, AbstractVector{<:AbstractString}} = "AllCountries",
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}} = "All"
    ) -> IntlServTrade

This dataset contains annual data on U.S. international trade in services.

Examples

julia> query = IntlServTrade(typeofservice = "AllTypesOfService", tradedirection = "Imports", areaorcountry = "Germany", year = 2014:2015)
IntlServTrade("AllTypesOfService", "Imports", "ALL", "Germany", "2014,2015")

julia> query = IntlServTrade(typeofservice = "Telecom", tradedirection = "Exports", affiliation = "UsParents")
IntlServTrade("Telecom", "Exports", "UsParents", "AllCountries", "ALL")

BEA.NIPA — Type

NIPA(
    tablename::AbstractString,
    frequency::Union{Frequency, AbstractVector{Frequency}},
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}}
    ) -> NIPA

This dataset contains data from the National Income and Product Accounts (NIPA) which include measures of the value and composition of U.S.production and the incomes generated in producing it.

Examples

julia> query = NIPA("T10101", ['A', 'Q'], "ALL")
NIPA("T10101", "A,Q", "ALL")

julia> query = NIPA("T20600", 'M', 2015:2016)
NIPA("T20600", "M", "2015,2016")

BEA.NIUnderlyingDetail — Type

NIUnderlyingDetail(
    tablename::AbstractString,
    frequency::Union{Frequency, AbstractVector{Frequency}},
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}}
    ) -> NIUnderlyingDetail

This dataset contains underlying detail data from the National Income and Product Accounts (NIPA) which include measures of the value and composition of U.S.production and the incomes generated in producing it.

Examples

julia> query = NIUnderlyingDetail("U20305", ['A', 'Q', 'M'], "ALL")
NIUnderlyingDetail("U20305", "A,M,Q", "ALL")

julia> query = NIUnderlyingDetail("U70205S", 'M', 2015:2016)
NIUnderlyingDetail("U70205S", "M", "2015,2016")

BEA.Regional — Type

Regional(
    tablename::AbstractString,
    linecode::Union{AbstractString, Integer},
    geofips::Union{AbstractString, AbstractVector{<:AbstractString}},
    year::Union{AbstractString, Integer, AbstractVector{<:Union{<:AbstractString, Integer}}} = "LAST5"
    ) -> Regional

The Regional dataset contains income and employment estimates from the Regional Economic Accounts by state, county, and metropolitan area

Examples

julia> query = Regional("CAINC1", 1, "County", 2012:2013)
Regional("CAINC1", "1", "County", "2012,2013")

julia> query = Regional("SAGDP9N", 2, "STATE", "ALL")
Regional("SAGDP9N", "2", "STATE", "ALL")

BEA.UnderlyingGDPbyIndustry — Type

UnderlyingGDPbyIndustry(
    tableid::Union{AbstractString, Integer, AbstractVector{<:Integer}},
    industry::Union{AbstractString, Integer, AbstractVector{<:Integer}},
    year::Union{AbstractString, Integer, AbstractVector{<:Integer}},
    frequency::Union{Char, AbstractVector{<:Char}} = 'A'
    ) -> UnderlyingGDPbyIndustry

The underlying GDP-by-industry dataset includes data in both current and chained (real) dollars. The dataset contains estimates for value added, gross output, and intermediate input statistics. This dataset is structurally similar to the GDPbyIndustry dataset (Appendix F), but contains additional industry detail.

Examples

julia> query = UnderlyingGDPbyIndustry(210, "ALL", 2010:2013)
UnderlyingGDPbyIndustry("210", "A", "2010,2011,2012,2013", "ALL")

julia> query = UnderlyingGDPbyIndustry("ALL", 11, 2012)
UnderlyingGDPbyIndustry("ALL", "A", "2012", "11")

Functions

BEA.bea_api_data — Method

bea_api_data(
    BEA_token::AbstractString,
    query::BEA_API_Datasets
    ) -> DataFrame

Return data from a query.

Examples

julia> query = FixedAssets("FAAt201", "ALL")
FixedAssets("FAAt201", "ALL")

julia> bea_api_data(BEA_token, query)
9888×10 DataFrame
  Row │ TableName  SeriesCode    LineNumber  LineDescription       TimePeriod  METRIC_NAME      CL_UNIT  UNIT_MULT  Da ⋯
      │ String     String        String      String                String      String           String   String     St ⋯
──────┼─────────────────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1925        Current Dollars  Level    9          22 ⋯
    2 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1926        Current Dollars  Level    9          23
    3 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1927        Current Dollars  Level    9          23
    4 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1928        Current Dollars  Level    9          24
    5 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1929        Current Dollars  Level    9          25 ⋯
    6 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1930        Current Dollars  Level    9          23
    7 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1931        Current Dollars  Level    9          20
    8 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1932        Current Dollars  Level    9          18
    9 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1933        Current Dollars  Level    9          19 ⋯
   10 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1934        Current Dollars  Level    9          19
   11 │ FAAt201    k1ptotl1es00  1           Private fixed assets  1935        Current Dollars  Level    9          19
  ⋮   │     ⋮           ⋮            ⋮                ⋮                ⋮              ⋮            ⋮         ⋮         ⋱
 9879 │ FAAt201    k1ntotl1ae50  103         Other                 2011        Current Dollars  Level    9          27
 9880 │ FAAt201    k1ntotl1ae50  103         Other                 2012        Current Dollars  Level    9          27 ⋯
 9881 │ FAAt201    k1ntotl1ae50  103         Other                 2013        Current Dollars  Level    9          27
 9882 │ FAAt201    k1ntotl1ae50  103         Other                 2014        Current Dollars  Level    9          28
 9883 │ FAAt201    k1ntotl1ae50  103         Other                 2015        Current Dollars  Level    9          28
 9884 │ FAAt201    k1ntotl1ae50  103         Other                 2016        Current Dollars  Level    9          28 ⋯
 9885 │ FAAt201    k1ntotl1ae50  103         Other                 2017        Current Dollars  Level    9          29
 9886 │ FAAt201    k1ntotl1ae50  103         Other                 2018        Current Dollars  Level    9          29
 9887 │ FAAt201    k1ntotl1ae50  103         Other                 2019        Current Dollars  Level    9          30
 9888 │ FAAt201    k1ntotl1ae50  103         Other                 2020        Current Dollars  Level    9          30 ⋯
                                                                                         2 columns and 9867 rows omitted

BEA.bea_api_datasets — Method

bea_api_datasets(BEA_token::AbstractString) -> DataFrame

Retrieves datasets currently available.

Examples

julia> bea_api_datasets(BEA_token)
12×2 DataFrame
 Row │ DatasetName              DatasetDescription
     │ String                   String
─────┼────────────────────────────────────────────────────────────
   1 │ NIPA                     Standard NIPA tables
   2 │ NIUnderlyingDetail       Standard NI underlying detail ta…
   3 │ MNE                      Multinational Enterprises
   4 │ FixedAssets              Standard Fixed Assets tables
   5 │ ITA                      International Transactions Accou…
   6 │ IIP                      International Investment Position
   7 │ InputOutput              Input-Output Data
   8 │ IntlServTrade            International Services Trade
   9 │ GDPbyIndustry            GDP by Industry
  10 │ Regional                 Regional data sets
  11 │ UnderlyingGDPbyIndustry  Underlying GDP by Industry
  12 │ APIDatasetMetaData       Metadata about other API datasets

BEA.bea_api_parameters — Method

bea_api_parameters(
    BEA_token::AbstractString,
    dataset::AbstractString
    ) -> DataFrame

Retrieves a list of the datasets currently offered.

Examples

julia> bea_api_parameters(BEA_token, "NIPA")
5×7 DataFrame
 Row │ ParameterName  ParameterDescription               ParameterDataType  ParameterIsRequiredFlag  ParameterDefaultV ⋯
     │ String         String                             String             Bool                     String?           ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ Frequency      A - Annual, Q-Quarterly, M-Month…  string                                true  missing           ⋯
   2 │ ShowMillions   A flag indicating that million-d…  string                               false  N
   3 │ TableID        The standard NIPA table identifi…  integer                              false  missing
   4 │ TableName      The new NIPA table identifier      string                               false  missing
   5 │ Year           List of year(s) of data to retri…  integer                               true  missing           ⋯
                                                                                                       3 columns omitted

BEA.bea_api_parametervalues — Method

bea_api_parametervalues(
    BEA_token::AbstractString,
    DatasetName::AbstractString,
    ParameterName::AbstractString;
    kwargs...) -> DataFrame

Retrieves a list of the valid values for a particular parameter.

Examples

julia> bea_api_parametervalues(BEA_token, "Regional", "TableName")
99×2 DataFrame
 Row │ Key       Desc
     │ String    String
─────┼─────────────────────────────────────────────
   1 │ CAEMP25N  Total Full-Time and Part-Time Em…
   2 │ CAEMP25S  Total Full-Time and Part-Time Em…
   3 │ CAGDP1    Gross Domestic Product (GDP) sum…
   4 │ CAGDP11   Contributions to percent change …
   5 │ CAGDP2    Gross domestic product (GDP) by …
   6 │ CAGDP8    Chain-type quantity indexes for …
   7 │ CAGDP9    Real GDP by county and metropoli…
   8 │ CAINC1    Personal Income Summary: Persona…
   9 │ CAINC30   Economic Profile (Non-Industry)
  10 │ CAINC35   Personal Current Transfer Receip…
  11 │ CAINC4    Personal Income and Employment b…
  ⋮  │    ⋮                      ⋮
  90 │ SQINC35   Personal Current Transfer Receip…
  91 │ SQINC4    Personal Income by Major Compone…
  92 │ SQINC5H   Personal Income by Major Compone…
  93 │ SQINC5N   Personal Income by Major Compone…
  94 │ SQINC5S   Personal Income by Major Compone…
  95 │ SQINC6N   Compensation of Employees by NAI…
  96 │ SQINC6S   Compensation of Employees by SIC…
  97 │ SQINC7H   Wages and Salaries by Industry (…
  98 │ SQINC7N   Wages and Salaries by NAICS Indu…
  99 │ SQINC7S   Wages and Salaries by SIC Indust…
                                    78 rows omitted

julia> bea_api_parametervalues(BEA_token, "MNE", "Country")
259×2 DataFrame
 Row │ key     desc
     │ String  String
─────┼──────────────────────────────
   1 │ all      all
   2 │ 000      all Countries Total
   3 │ 600     Afghanistan
   4 │ 499     Africa
   5 │ 350     Albania
   6 │ 400     Algeria
   7 │ 300     Andorra
   8 │ 401     Angola
   9 │ 272     Anguilla
  10 │ 273     Antigua and Barbuda
  11 │ 200     Argentina
  ⋮  │   ⋮              ⋮
 250 │ 653     Vietnam
 251 │ 440     Western Sahara
 252 │ 637     Western Samoa
 253 │ 514     Yemen
 254 │ 500     Yemen (Aden)
 255 │ 514     Yemen (Sanaa)
 256 │ 328     Yugoslavia
 257 │ 408     Zaire
 258 │ 448     Zambia
 259 │ 431     Zimbabwe
                    238 rows omitted

julia> bea_api_parametervalues(BEA_token, "Regional", "LineCode", TableName = "SQINC7N", Year = 2020)
29×2 DataFrame
 Row │ Key     Desc
     │ String  String
─────┼───────────────────────────────────────────
   1 │ 100     [SQINC7N] Private wages and sala…
   2 │ 1000    [SQINC7N] Private nonfarm wages …
   3 │ 1100    [SQINC7N] Private nonfarm wages …
   4 │ 1200    [SQINC7N] Private nonfarm wages …
   5 │ 1300    [SQINC7N] Private nonfarm wages …
   6 │ 1400    [SQINC7N] Private nonfarm wages …
   7 │ 1500    [SQINC7N] Private nonfarm wages …
   8 │ 1600    [SQINC7N] Private nonfarm wages …
   9 │ 1700    [SQINC7N] Private nonfarm wages …
  10 │ 1800    [SQINC7N] Private nonfarm wages …
  11 │ 1900    [SQINC7N] Private nonfarm wages …
  ⋮  │   ⋮                     ⋮
  20 │ 500     [SQINC7N] Private nonfarm wages …
  21 │ 510     [SQINC7N] Private nonfarm wages …
  22 │ 530     [SQINC7N] Private nonfarm wages …
  23 │ 600     [SQINC7N] Private nonfarm wages …
  24 │ 700     [SQINC7N] Private nonfarm wages …
  25 │ 800     [SQINC7N] Private nonfarm wages …
  26 │ 81      [SQINC7N] Farm wages and salarie…
  27 │ 82      [SQINC7N] Nonfarm wages and sala…
  28 │ 90      [SQINC7N] Private nonfarm wages …
  29 │ 900     [SQINC7N] Private nonfarm wages …
                                   8 rows omitted